move raw_monitor and transect_minotor to operations/
